refactor: 트랜잭션 분리
